refactor: remove standalone CSS outputs
Move feature SCSS partials from `core/styles/`
to core/styles/features/`.

BREAKING CHANGE: The following CSS files are no longer published
 as separate build outputs: `a11y.css`, `animation.css`, `badge.css`,
 `core.css`, `disable-animation.css`, `layout.css`, `lists.css`,
 `normalize.css`, `scrollbar.css`, `table.css`, `typography.css`

If you were importing any of these files directly and need only a
subset of styles, create a custom Sass entry file and `@use` the
relevant partials from `core/styles/features/` instead.
