+++ src/elements/core/styles/features/a11y.scss
@@ -1,5 +1,5 @@
-@use './mixins/a11y';
-@use './internals/screen-reader-only';
+@use '../mixins/a11y';
+@use '../internals/screen-reader-only';
 
 :is(.sbb-focus-outline, .sbb-focus-outline-dark):focus-visible {
   @include a11y.focus-outline;
