+++ src/elements/core/styles/features/animation.scss
@@ -1,4 +1,4 @@
-@use './mixins/animation';
+@use '../mixins/animation';
 
 .sbb-disable-animation {
   @include animation.disable-animation;
