+++ src/elements/core/styles/features/badge.scss
@@ -1,3 +1,3 @@
-@use './mixins/badge';
+@use '../mixins/badge';
 
 @include badge.badge-attribute;
