+++ src/elements/core/styles/features/layout.private.scss
@@ -1,5 +1,5 @@
-@use './mixins/dev_and_debug';
-@use './core/mediaqueries';
+@use '../mixins/dev_and_debug';
+@use '../core/mediaqueries';
 @use '@sbb-esta/lyne-design-tokens/dist/scss/sbb-variables.scss' as sbb-tokens;
 
 .visualized-page-spacing div,
