+++ src/elements/core/styles/features/layout.scss
@@ -1,4 +1,4 @@
-@use './mixins/layout';
+@use '../mixins/layout';
 
 .sbb-grid,
 .sbb-page-spacing {
