+++ src/elements/core/styles/features/lists.scss
@@ -1,4 +1,4 @@
-@use './mixins/lists';
+@use '../mixins/lists';
 
 .sbb-list {
   @include lists.list;
