+++ src/elements/core/styles/features/scrollbar.scss
@@ -1,4 +1,4 @@
-@use './mixins/scrollbar';
+@use '../mixins/scrollbar';
 
 :is(
   .sbb-scrollbar,
