+++ src/elements/core/styles/features/typography.private.scss
@@ -1,4 +1,4 @@
-@use './mixins/link';
+@use '../mixins/link';
 
 .sbb-link {
   @include link.link;
