+++ src/elements/core/styles/features/typography.scss
@@ -1,4 +1,4 @@
-@use './mixins/typo';
+@use '../mixins/typo';
 
 sub {
   @include typo.sub;
