+++ src/elements/core/styles/theme.scss
@@ -2,18 +2,15 @@ $theme: 'standard' !default;
 
 @use './_index' as sbb;
 
-@use './normalize';
-@use './core' with (
+@use 'features/normalize';
+@use 'features/core' with (
   $theme: $theme
 );
-@use './typography';
-@use './a11y';
-@use './badge';
-@use './animation';
-@use './image';
-@use './layout';
-@use './lists';
-@use './scrollbar';
-@use '../../table/table.global' with (
-  $theme: $theme
-);
+@use 'features/typography';
+@use 'features/a11y';
+@use 'features/badge';
+@use 'features/animation';
+@use 'features/image';
+@use 'features/layout';
+@use 'features/lists';
+@use 'features/scrollbar';
