+++ src/storybook/guides/theming.mdx
@@ -68,7 +68,7 @@ In certain safety relevant applications, it might be necessary to change the cur
 ### Exposed CSS files
 
 Basically, all our styles are included in the theme files (e.g. `standard-theme.css`) which should be included in your application.
-However, if you would like to more specifically pick what you need, consider the following CSS files available.
+However, if you would like to more specifically pick what you need, create an own Sass file importing the needed features from our Sass API.
 
 | File name                       | Description                                                                                           |
 | ------------------------------- | ----------------------------------------------------------------------------------------------------- |
@@ -80,20 +80,6 @@ However, if you would like to more specifically pick what you need, consider the
 | `lean-safety-theme.css`         | Contains the lean theme with primary and brand colors set to an alternative color (gray).             |
 |                                 |                                                                                                       |
 | `font-characters-extension.css` | Provides full character set of SBB fonts, needs larger files to load.                                 |
-|                                 |                                                                                                       |
-| `normalize.css`                 | Contains general browser resetting styles which can be useful for your application.                   |
-|                                 |                                                                                                       |
-| `core.css`                      | Contains mandatory basics to use lyne-components (including design tokens).                           |
-|                                 |                                                                                                       |
-| `a11y.css`                      | Provides accessibility related CSS classes.                                                           |
-| `animation.css`                 | Provides CSS classes to disable animation (e.g. for testing) on any level.                            |
-| `badge.css`                     | Provides badge styling.                                                                               |
-| `disable-animation.css`         | Disables animations on root level, ideally for a global test setup.                                   |
-| `layout.css`                    | Provides layout related CSS classes (e.g. page spacing, grid).                                        |
-| `lists.css`                     | Provides CSS classes to style lists.                                                                  |
-| `scrollbar.css`                 | Provides CSS classes to style a scrollbar.                                                            |
-| `table.css`                     | Provides CSS classes to style a table.                                                                |
-| `typography.css`                | Provides typography related CSS classes.                                                              |
 
 ### Full Font
